Добавляет реализацию View паттерна MVC для экранов приветствия и игровых экранов
